feat: add `BigDecimal` and `BigInteger` serializers with tests
- Implement `BigDecimalSerializer` and `BigIntegerSerializer` for handling `BigDecimal` and `BigInteger` types.
- Add unit tests to validate serialization and deserialization for both serializers, ensuring accurate YAML output for various cases.

+++ src/main/kotlin/dev/s7a/ktconfig/serializer/NumberSerializer.kt
@@ -23,7 +23,7 @@ object NumberSerializer : Serializer.Keyable<Number> {
                     INFINITY -> Double.POSITIVE_INFINITY
                     NEGATIVE_INFINITY -> Double.NEGATIVE_INFINITY
                     NOT_A_NUMBER -> Double.NaN
-                    else -> BigDecimal(value).toDouble()
+                    else -> BigDecimal(value)
                 }
             }
             else -> throw IllegalArgumentException("Cannot convert to Number: ${value::class.simpleName}")
